Docstring fixes
