+++ cca_zoo/models/_partialcca.py
@@ -1,7 +1,6 @@
 from typing import Iterable, Union
 
 import numpy as np
-from scipy.linalg import block_diag
 from sklearn.utils.validation import check_is_fitted
 
 from cca_zoo.models import MCCA
