+++ cca_zoo/models/_prcca.py
@@ -5,7 +5,6 @@ import numpy as np
 from scipy.linalg import block_diag
 
 from cca_zoo.models._multiview._mcca import MCCA
-from ..utils import _process_parameter
 
 
 class PRCCA(MCCA):
@@ -55,9 +54,7 @@ class PRCCA(MCCA):
         views = self._validate_inputs(views)
         self._check_params()
         views = self._pretransform(views, idxs)
-        views, C, D = self._setup_evp(views, idxs, **kwargs)
-        self._solve_evp(views, C, D, **kwargs)
-        return self
+        return super().fit(views, idxs=idxs)
 
     def _pretransform(self, views, idxs):
         X_1 = [view[:, idx] for view, idx in zip(views, idxs)]
@@ -71,7 +68,7 @@ class PRCCA(MCCA):
             for X_1, X_2 in zip(X_1, X_2)]
         return views
 
-    def _setup_evp(self, views: Iterable[np.ndarray], idxs, **kwargs):
+    def _setup_evp(self, views: Iterable[np.ndarray], idxs=None, **kwargs):
         all_views = np.concatenate(views, axis=1)
         C = all_views.T @ all_views / self.n
         penalties = [np.zeros((view.shape[1])) for view in views]
@@ -84,118 +81,4 @@ class PRCCA(MCCA):
         D_smallest_eig = min(0, np.linalg.eigvalsh(D).min()) - self.eps
         D = D - D_smallest_eig * np.eye(D.shape[0])
         self.splits = np.cumsum([0] + [view.shape[1] for view in views])
-        return views, C, D
-
-
-class GRCCA(PRCCA):
-    """
-    Grouped Regularized Canonical Correlation Analysis
-
-    References
-    ----------
-    Tuzhilina, Elena, Leonardo Tozzi, and Trevor Hastie. "Canonical correlation analysis in high dimensions with structured regularization." Statistical Modelling (2021): 1471082X211041033.
-    """
-
-    def __init__(
-            self,
-            latent_dims: int = 1,
-            scale: bool = True,
-            centre=True,
-            copy_data=True,
-            random_state=None,
-            eps=1e-3,
-            c: float = 0,
-            mu: float = 0,
-    ):
-        super().__init__(
-            latent_dims=latent_dims,
-            scale=scale,
-            centre=centre,
-            copy_data=copy_data,
-            random_state=random_state,
-            eps=eps,
-            c=c,
-        )
-        self.mu = mu
-
-    def _check_params(self):
-        self.mu = _process_parameter("mu", self.mu, 0, self.n_views)
-        self.c = _process_parameter("c", self.c, 0, self.n_views)
-
-    def fit(self, views: Iterable[np.ndarray], y=None, subject_groups=None, **kwargs):
-        """
-        Parameters
-        ----------
-        views : list/tuple of numpy arrays or array likes with the same number of rows (samples)
-        y : None
-        subject_groups : list/tuple of integer numpy arrays or array likes with dimensions (samples,)
-        kwargs: any additional keyword arguments required by the given model
-
-        """
-        if subject_groups is None:
-            warnings.warn(f"No feature groups provided, using all features")
-            subject_groups = [np.ones(views[0].shape[0], dtype=int)] * len(views)
-        for subject_group in subject_groups:
-            assert subject_group.dtype == int, "subject_groups must be integers"
-        views = self._validate_inputs(views)
-        self._check_params()
-        views, idxs = self._preprocess(views, subject_groups)
-        self.weights = (
-            PRCCA(latent_dims=self.latent_dims, scale=False, centre=False, c=1)
-            .fit(views, idxs=idxs, **kwargs)
-            .weights
-        )
-        self._transform_weights(views, subject_groups)
-        return self
-
-    def _preprocess(self, views, subject_groups):
-        views, idxs = list(
-            zip(
-                *[
-                    self._process_view(view, group, mu, c)
-                    for view, group, mu, c in zip(views, subject_groups, self.mu, self.c)
-                ]
-            )
-        )
-        return views, idxs
-
-    @staticmethod
-    def _process_view(view, group, mu, c):
-        if c > 0:
-            ids, unique_inverse, unique_counts, group_means = _group_mean(view, group)
-            if mu == 0:
-                mu = 1
-                idx = view.shape[1] - 1
-            else:
-                idx = view.shape[1] + group_means.shape[1] - 1
-            view_1 = (view - group_means[:, unique_inverse]) / c
-            view_2 = group_means / np.sqrt(mu / unique_counts)
-            return np.hstack((view_1, view_2)), idx
-        else:
-            return view, view.shape[1] - 1
-
-    def _transform_weights(self, views, groups):
-        for i, (view, group) in enumerate(zip(views, groups)):
-            if self.c[i] > 0:
-                weights_1 = self.weights[i][: len(group)]
-                weights_2 = self.weights[i][len(group):]
-                ids, unique_inverse, unique_counts, group_means = _group_mean(
-                    weights_1.T, group
-                )
-                weights_1 = (weights_1 - group_means[:, unique_inverse].T) / self.c[i]
-                if self.mu[i] == 0:
-                    mu = 1
-                else:
-                    mu = self.mu[i]
-                weights_2 = weights_2 / np.sqrt(
-                    mu * np.expand_dims(unique_counts, axis=1)
-                )
-                self.weights[i] = weights_1 + weights_2[group]
-
-
-def _group_mean(view, group):
-    ids, unique_inverse, unique_counts = np.unique(
-        group, return_inverse=True, return_counts=True
-    )
-    group_means = np.array([view[:, group == id].mean(axis=1) for id in ids]).T
-    return ids, unique_inverse, unique_counts, group_means
+        return C, D
